better install.sh

+++ install.sh
@@ -13,8 +13,19 @@ sleep 1
 echo ""
 echo "Where would you like to install the Honeypot? ( ex: /var/www )"
 read directory
+if [ -z "$directory" ]; then
+    directory="/var/www"
+fi
 echo ""
 echo "Installing honeypot to: $directory/phpmyadmin/"
+echo ""
+echo "What web user does your server use? ( ex: apache, www-data )"
+read webUser
+if [ -z "$webUser" ]; then
+    webUser="apache"
+fi
+echo ""
+echo "All files will be for: $webUser:$webUser"
 echo ""
 echo "What Username would you like to use for PhpMyAdmin? (default: root)"
 read phpmyadminUsername
@@ -29,7 +40,7 @@ fi
 echo ""
 echo "PhpMyAdmin Username and Password Configured ($phpmyadminUsername/$phpmyadminPassword)"
 echo ""
-mv ./phpmyadmin-interactive $directory/phpmyadmin
+cp -r ./phpmyadmin-interactive $directory/phpmyadmin
 echo "Rename the honeypot log filename (no extension and make this random): "
 read phpmyadminLog
 if [ -z "$phpmyadminLog" ]; then
@@ -45,13 +56,22 @@ sed -i "/\$File = \"log.txt\";/c\\\$File = \"$phpmyadminLog.txt\";" $directory/p
 sed -i "/\$myFile = \"log.txt\";/c\\\$myFile = \"$phpmyadminLog.txt\";" $directory/phpmyadmin/login.php
 sed -i 's/log.txt/'$phpmyadminLog.txt'/g' $directory/phpmyadmin/master-config/index.php
 sed -i 's/log.txt/'$phpmyadminLog.txt'/g' $directory/phpmyadmin/master-config/phpinfo.php
-mv $directory/phpmyadmin/log.txt $directory/phpmyadmin/$phpmyadminLog.txt
+if [ -f "$directory/phpmyadmin/log.txt" ]; then
+    mv $directory/phpmyadmin/log.txt $directory/phpmyadmin/$phpmyadminLog.txt
+else
+    touch $directory/phpmyadmin/$phpmyadminLog.txt
+fi
 sed -i 's/USERNAME/'$phpmyadminUsername'/g' $directory/phpmyadmin/login.php
 sed -i 's/PASSWORD/'$phpmyadminPassword'/g' $directory/phpmyadmin/login.php
 if [ -f ../phpmyadmin_honeypot/README.md ]; then
-     rm -rf ../phpmyadmin_honeypot/
+    echo "Delete phpmyadmin_honeypot folder? (Y/N):"
+    read deleteFolder
+    if [[ $deleteFolder == [yY] ]]; then
+        rm -rf ../phpmyadmin_honeypot/
+    fi
 fi
 echo ""
+sudo chown -R $webUser:$webUser $directory/phpmyadmin/
 echo "Fake PhpMyAdmin has been configured successfully and can be accessed via http://127.0.0.1/phpmyadmin/"
 echo "There are also one fake phpinfo page: $directory/phpmyadmin/phpinfo.php"
 echo "Logs will write to: $directory/phpmyadmin/$phpmyadminLog.txt"
